Move ray module into backend module
