+++ crates/bevy_picking_core/src/backend.rs
@@ -21,6 +21,12 @@
 //!   use it for optimization purposes. For example, a backend that traverses a spatial hierarchy
 //!   may want to early exit if it intersects an entity that blocks lower entities from being
 //!   picked.
+//!
+//! ### Raycasting Backends
+//!
+//! Backends that require a ray to cast into the scene should use [`ray::RayMap`]. This
+//! automatically constructs rays in world space for all cameras and pointers, handling details like
+//! viewports and DPI for you.
 
 use bevy_ecs::prelude::*;
 use bevy_math::Vec3;
@@ -28,7 +34,7 @@ use bevy_reflect::Reflect;
 
 /// Common imports for implementing a picking backend.
 pub mod prelude {
-    pub use super::{HitData, PointerHits};
+    pub use super::{ray::RayMap, HitData, PointerHits};
     pub use crate::{
         pointer::{PointerId, PointerLocation},
         PickSet, Pickable,
@@ -105,5 +111,116 @@ impl HitData {
             position,
             normal,
         }
+    }
+}
+
+pub mod ray {
+    //! Types and systems for constructing rays from cameras and pointers.
+
+    use crate::backend::prelude::{PointerId, PointerLocation};
+    use bevy_ecs::prelude::*;
+    use bevy_math::Ray;
+    use bevy_reflect::Reflect;
+    use bevy_render::camera::Camera;
+    use bevy_transform::prelude::GlobalTransform;
+    use bevy_utils::{hashbrown::hash_map::Iter, HashMap};
+    use bevy_window::PrimaryWindow;
+
+    /// Identifies a ray constructed from some (pointer, camera) combination. A pointer can be over
+    /// multiple cameras, which is why a single pointer may have multiple rays.
+    #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, Reflect)]
+    pub struct RayId {
+        /// The camera whose projection was used to calculate the ray.
+        pub camera: Entity,
+        /// The pointer whose pixel coordinates were used to calculate the ray.
+        pub pointer: PointerId,
+    }
+
+    impl RayId {
+        /// Construct a [`RayId`].
+        pub fn new(camera: Entity, pointer: PointerId) -> Self {
+            Self { camera, pointer }
+        }
+    }
+
+    /// A map from [`RayId`] to [`Ray`].
+    ///
+    /// This map is cleared and re-populated every frame before any backends run. Ray-based picking
+    /// backends should use this when possible, as it automatically handles viewports, DPI, and
+    /// other details of building rays from pointer locations.
+    ///
+    /// ## Usage
+    ///
+    /// Iterate over each [`Ray`] and its [`RayId`] with [`RayMap::iter`].
+    ///
+    /// ```
+    /// // My raycasting backend
+    /// pub fn update_hits(ray_map: Res<RayMap>, mut output_events: EventWriter<PointerHits>,) {
+    ///     for (&ray_id, &ray) in ray_map.iter() {
+    ///         /// Run a raycast with each ray, returning any `PointerHits` found.
+    ///     }
+    /// }
+    /// ```
+    #[derive(Clone, Debug, Default, Resource)]
+    pub struct RayMap {
+        map: HashMap<RayId, Ray>,
+    }
+
+    impl RayMap {
+        /// Iterates over all world space rays for every picking pointer.
+        pub fn iter(&self) -> Iter<'_, RayId, Ray> {
+            self.map.iter()
+        }
+
+        /// The hash map of all rays cast in the current frame.
+        pub fn map(&self) -> &HashMap<RayId, Ray> {
+            &self.map
+        }
+
+        /// Clears the [`RayMap`] and re-populates it with one ray for each
+        /// combination of pointer entity and camera entity where the pointer
+        /// intersects the camera's viewport.
+        pub fn repopulate(
+            mut ray_map: ResMut<Self>,
+            primary_window_entity: Query<Entity, With<PrimaryWindow>>,
+            cameras: Query<(Entity, &Camera, &GlobalTransform)>,
+            pointers: Query<(&PointerId, &PointerLocation)>,
+        ) {
+            ray_map.map.clear();
+
+            for (camera_entity, camera, camera_tfm) in &cameras {
+                if !camera.is_active {
+                    continue;
+                }
+
+                for (&pointer_id, pointer_loc) in &pointers {
+                    if let Some(ray) =
+                        make_ray(&primary_window_entity, camera, camera_tfm, pointer_loc)
+                    {
+                        ray_map
+                            .map
+                            .insert(RayId::new(camera_entity, pointer_id), ray);
+                    }
+                }
+            }
+        }
+    }
+
+    fn make_ray(
+        primary_window_entity: &Query<Entity, With<PrimaryWindow>>,
+        camera: &Camera,
+        camera_tfm: &GlobalTransform,
+        pointer_loc: &PointerLocation,
+    ) -> Option<Ray> {
+        let pointer_loc = pointer_loc.location()?;
+        if !pointer_loc.is_in_viewport(camera, primary_window_entity) {
+            return None;
+        }
+        let mut viewport_pos = pointer_loc.position;
+        if let Some(viewport) = &camera.viewport {
+            let viewport_logical = camera.to_logical(viewport.physical_position)?;
+            viewport_pos -= viewport_logical;
+        }
+        camera.viewport_to_world(camera_tfm, viewport_pos)
     }
 }
